Move/merge impls out of `ty/mod.rs`
